Break out expression resolution
