+++ crates/expressions/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "spin-variables"
+name = "spin-expressions"
 version = { workspace = true }
 authors = { workspace = true }
 edition = { workspace = true }
@@ -10,9 +10,6 @@ async-trait = "0.1"
 dotenvy = "0.15"
 once_cell = "1"
 spin-app = { path = "../app" }
-spin-core = { path = "../core" }
-spin-expressions = { path = "../expressions" }
-spin-world = { path = "../world" }
 thiserror = "1"
 tokio = { version = "1", features = ["rt-multi-thread"] }
 vaultrs = "0.6.2"
